backend/pyproject.toml
[project]
name = "starlink-backend"
version = "0.1.0"
description = "Starlink Chat Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.5.0",
    "PyJWT[crypto]>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
    "langchain>=1.0.7",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-openai>=1.0.3",
    "langfuse>=3.10.0,<3.12.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "exceptiongroup>=1.0.0",
    "openpyxl>=3.1.0",
    "deepagents>=0.4.12",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
